+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "hatchling.build"
 [project]
 name = "SOAPify-GMPavanLab"
 description = "A package for creating and studying SOAP fingerprints"
-readme = { "file" = "Readme.md", "content-type" = "text/markdown" }
+readme = "Readme.md"
 requires-python = ">=3.8, <3.10"
 license = "MIT"
 keywords = ["SOAP", "Molecular Dynamics", "simulation", "simulation analysis"]
@@ -81,6 +81,8 @@ extra-dependencies = [
 [tool.hatch.envs.docs.scripts]
 cov = 'cd docs && make coverage'
 build = 'cd docs && make html'
+clean = 'cd docs && make clean'
+cleanbuild = 'cd docs && make clean build'
 
 [tool.hatch.envs.test]
 template="SOAP"
